chore: update Uno.Sdk to version 6.4.0-dev.32
The `global.json` file has been updated to change the version
of the `Uno.Sdk` in the `msbuild-sdks` section from
`6.4.0-dev.17` to `6.4.0-dev.32`. This reflects an upgrade
to a newer development version of the SDK.

The `allowPrerelease` property in the `sdk` section remains
unchanged and is still set to `false`.